updated main, addressbook and clean_folder
Виніс clean_folder до основного проєкту і трішки вніс правки в нього, додав адресу в addressbook, і додав запуск clean_folder в меню
