--- Py6PersonalAssistant/clean_folder/file_parser.py
@@ -1,127 +1,0 @@
-import sys
-from pathlib import Path
-
-#list for images
-JPEG_IMAGES = []
-JPG_IMAGES = []
-PNG_IMAGES = []
-SVG_IMAGES = []
-#list for audio
-MP3_AUDIO = []
-OGG_AUDIO = []
-WAV_AUDIO = []
-AMR_AUDIO = []
-#list for video
-MP4_VIDEO = []
-AVI_VIDEO = []
-MOV_VIDEO = []
-MKV_VIDEO = []
-#list for doc
-DOC_DOC = []
-DOCX_DOC = []
-TXT_DOC = []
-PDF_DOC = []
-XLSX_DOC = []
-PPTX_DOC = []
-#list for ARCHIVES
-ARCHIVES = []
-GZ_ARCHIVES = []
-TAR_ARCHIVES = []
-
-OTHER = []
-
-REGISTER_EXTENSIONS = {
-    'JPEG': JPEG_IMAGES,
-    'PNG': PNG_IMAGES,
-    'JPG': JPG_IMAGES,
-    'SVG': SVG_IMAGES,
-    'MP3': MP3_AUDIO,
-    'OGG': OGG_AUDIO,
-    'WAV': WAV_AUDIO,
-    'AMR': AMR_AUDIO,
-    'MP4': MP4_VIDEO,
-    'AVI': AVI_VIDEO,
-    'MOV': MOV_VIDEO,
-    'MKV': MKV_VIDEO,
-    'DOC': DOC_DOC,
-    'DOCX': DOCX_DOC,
-    'TXT': TXT_DOC,
-    'PDF': PDF_DOC,
-    'XLSX': XLSX_DOC,
-    'PPTX': PPTX_DOC,
-    'ZIP': ARCHIVES,
-    'GZ': GZ_ARCHIVES,
-    'TAR': TAR_ARCHIVES 
-}
-
-FOLDERS = []
-EXTENSIONS = set()
-UNKNOWN = set()
-
-
-def get_extension(filename: str) -> str:
-    # превращаем расширение файла в название папки .jpg -> JPG
-    return Path(filename).suffix[1:].upper()
-
-
-def scan(folder: Path) -> None:
-    for item in folder.iterdir():
-        # Если это папка то добавляем ее с список FOLDERS и преходим к следующему элементу папки
-        if item.is_dir():
-            # проверяем, чтобы папка не была той в которую мы складываем уже файлы
-            if item.name not in ('archives', 'video', 'audio', 'documents', 'images', 'OTHER'):
-                FOLDERS.append(item)
-                #  сканируем эту вложенную папку - рекурсия
-                scan(item)
-            #  перейти к следующему элементу в сканируемой папке
-            continue
-
-        #  Пошла работа с файлом
-        ext = get_extension(item.name)  # взять расширение
-        fullname = folder / item.name  # взять полный путь к файлу
-        if not ext:  # если у файла нет расширения добавить к неизвестным
-            OTHER.append(fullname)
-        else:
-            try:
-                # взять список куда положить полный путь к файлу
-                container = REGISTER_EXTENSIONS[ext]
-                EXTENSIONS.add(ext)
-                container.append(fullname)
-            except KeyError:
-                # Если мы не регистрировали расширение в REGISTER_EXTENSIONS, то добавить в другое
-                UNKNOWN.add(ext)
-                OTHER.append(fullname)
-
-
-if __name__ == '__main__':
-    folder_for_scan = sys.argv[1]
-    print(f'Start in folder {folder_for_scan}')
-#Print all rezults!
-    scan(Path(folder_for_scan))
-    print(f'Images jpeg: {JPEG_IMAGES}')
-    print(f'Images jpg: {JPG_IMAGES}')
-    print(f'Images png: {PNG_IMAGES}')
-    print(f'Images svg: {SVG_IMAGES}')
-    print(f'Audio mp3: {MP3_AUDIO}')
-    print(f'Audio ogg: {OGG_AUDIO}')
-    print(f'Audio wav: {WAV_AUDIO}')
-    print(f'Audio amr: {AMR_AUDIO}')
-    print(f'Video mp4: {MP4_VIDEO}')
-    print(f'Video avi: {AVI_VIDEO}')
-    print(f'Video mov: {MOV_VIDEO}')
-    print(f'Video mkv: {MKV_VIDEO}')
-    print(f'Documents doc: {DOC_DOC}')
-
-# DOCX_DOC 
-# TXT_DOC 
-# PDF_DOC 
-# XLSX_DOC 
-# PPTX_DOC
-
-    print(f'Archives: {ARCHIVES}')
-    print(f'Archives: {GZ_ARCHIVES}')
-    print(f'Archives: {TAR_ARCHIVES}')
-    print(f'Types of files in folder: {EXTENSIONS}')
-    print(f'Unknown files of types: {UNKNOWN}')
-
-    print(FOLDERS[::-1])
