+++ Py6PersonalAssistant/clean_folder_tools/file_parser.py
@@ -1,0 +1,127 @@
+import sys
+from pathlib import Path
+
+JPEG_IMAGES = []
+JPG_IMAGES = []
+PNG_IMAGES = []
+SVG_IMAGES = []
+
+AVI_VIDEO = []
+MP4_VIDEO = []
+MOV_VIDEO = []
+MKV_VIDEO = []
+
+DOC_DOCUMENTS = []
+DOCX_DOCUMENTS = []
+TXT_DOCUMENTS = []
+PDF_DOCUMENTS = []
+XLSX_DOCUMENTS = []
+PPTX_DOCUMENTS = []
+
+MP3_AUDIO = []
+OGG_AUDIO = []
+WAV_AUDIO = []
+AMR_AUDIO = []
+
+ZIP_ARCHIVES = []
+GZ_ARCHIVES = []
+TAR_ARCHIVES = []
+
+OTHER = []
+
+REGISTER_EXTENSIONS = {
+
+    'JPEG': JPEG_IMAGES,
+    'PNG': PNG_IMAGES,
+    'JPG': JPG_IMAGES,
+    'SVG': SVG_IMAGES,
+
+    'AVI': AVI_VIDEO,
+    'MP4': MP4_VIDEO,
+    'MOV': MOV_VIDEO,
+    'MKV': MKV_VIDEO,
+
+    'DOC': DOC_DOCUMENTS,
+    'DOCX': DOCX_DOCUMENTS,
+    'TXT': TXT_DOCUMENTS,
+    'PDF': PDF_DOCUMENTS,
+    'XLSX': XLSX_DOCUMENTS,
+    'PPTX': PPTX_DOCUMENTS,
+
+    'MP3': MP3_AUDIO,
+    'OGG': OGG_AUDIO,
+    'WAV': WAV_AUDIO,
+    'AMR': AMR_AUDIO,
+
+    'ZIP': ZIP_ARCHIVES,
+    'GZ': GZ_ARCHIVES,
+    'TAR': TAR_ARCHIVES,
+}
+
+FOLDERS = []
+EXTENSIONS = set()
+UNKNOWN = set()
+
+
+def get_extension(filename: str) -> str:
+    return Path(filename).suffix[1:].upper()
+
+
+def scan(folder: Path) -> None:
+    for item in folder.iterdir():
+        if item.is_dir():
+            if item.name not in ('archives', 'video', 'audio', 'documents', 'images', 'OTHER'):
+                FOLDERS.append(item)
+                scan(item)
+            continue
+
+        ext = get_extension(item.name)
+        fullname = folder / item.name
+        if not ext:
+            OTHER.append(fullname)
+        else:
+            try:
+                container = REGISTER_EXTENSIONS[ext]
+                EXTENSIONS.add(ext)
+                container.append(fullname)
+            except KeyError:
+                UNKNOWN.add(ext)
+                OTHER.append(fullname)
+
+
+if __name__ == '__main__':
+    folder_for_scan = sys.argv[1]
+    print(f'Start in folder {folder_for_scan}')
+
+    scan(Path(folder_for_scan))
+
+    print(f'Images jpeg: {JPEG_IMAGES}')
+    print(f'Images jpg: {JPG_IMAGES}')
+    print(f'Images svg: {SVG_IMAGES}')
+    print(f'Images png: {PNG_IMAGES}')
+
+    print(f'Video avi: {AVI_VIDEO}')
+    print(f'Video mp4: {MP4_VIDEO}')
+    print(f'Video mov: {MOV_VIDEO}')
+    print(f'Video mkv: {MKV_VIDEO}')
+
+    print(f'Docs doc: {DOC_DOCUMENTS}')
+    print(f'Docs docx: {DOCX_DOCUMENTS}')
+    print(f'Docs txt: {TXT_DOCUMENTS}')
+    print(f'Docs pdf: {PDF_DOCUMENTS}')
+    print(f'Docs xlsx: {XLSX_DOCUMENTS}')
+    print(f'Docs pptx: {PPTX_DOCUMENTS}')
+
+    print(f'Audio mp3: {MP3_AUDIO}')
+    print(f'Audio ogg: {OGG_AUDIO}')
+    print(f'Audio wav: {WAV_AUDIO}')
+    print(f'Audio amr: {AMR_AUDIO}')
+
+    print(f'Archives zip: {ZIP_ARCHIVES}')
+    print(f'Archives gz: {GZ_ARCHIVES}')
+    print(f'Archives tar: {TAR_ARCHIVES}')
+
+    print(f'Types of files in folder: {EXTENSIONS}')
+    print(f'Unknown files of types: {UNKNOWN}')
+
+    print(FOLDERS[::-1])
