+++ Py6PersonalAssistant/clean_folder_tools/normalize.py
@@ -2,7 +2,7 @@ import re
 
 CYRILLIC_SYMBOLS = 'абвгдеёжзийклмнопрстуфхцчшщъыьэюяєіїґ'
 TRANSLATION = ("a", "b", "v", "g", "d", "e", "e", "j", "z", "i", "j", "k", "l", "m", "n", "o", "p", "r", "s", "t", "u",
-               "f", "h", "ts", "ch", "sh", "sch", "", "y", "", "e", "yu", "u", "ja", "je", "ji", "g")
+               "f", "h", "ts", "ch", "sh", "sch", "", "y", "", "e", "yu", "ja", "je", "i", "ji", "g")
 
 TRANS = {}
 for c, l in zip(CYRILLIC_SYMBOLS, TRANSLATION):
@@ -12,5 +12,5 @@ for c, l in zip(CYRILLIC_SYMBOLS, TRANSLATION):
 
 def normalize(name: str) -> str:
     t_name = name.translate(TRANS)
-    t_name = re.sub(r'\W', '_', t_name)
+    t_name = re.sub(r'^\W{2}', '_', t_name)
     return t_name
